Differentiate between Background and Font colors
Right now, we have a Color and a FontColor. We want to be able to set a
different font color and background color for all the text/slides, where also
the generic one has an impact.

To do so, we rename Color to BackgroundColor and we make FontColor available in
more sections: this way, it is easier to understand what to set instead of
relying on the code to change the appropriate color when in a "slide" or in a "text"