Refactor parser
This commit refactors the parser quite a lot.
There is now a VecDeque instead of a Vec to hold the tokens, and we pop from
the front of the queue instead of bookkeeping the index. This way we have less
error due to indexing, since we pop the same amount of element we consume.

Moreover, we have now a specific structure to store the slide to be built,
which is better than going to take the slide from the ones that have been built
already and append to that.

Unwraps are not removed yet in the parser.