First implementation of the crossterm backend
Many TODOs and many unwrap, but the main functionality is there.
Note that it is quite annoying to use since we need to manually set the
terminal font size, but I don't see an easy way to get around this...

Maybe the sdl backend is the only one that does really make sense?